Transfer ownership of ephemeral znodes.
If the lighthouse-reporter process is restarted, a new zookeeper session
made.  This new session will interact with the old session's ephemeral
znodes when calling report_up() *but* usually after a timeout the node
is deleted as the old session is gone and the reporter is
none-the-wiser.

Fix is to delete and recreate the znode in a transaction in order to
"transfer" ownership to the new session.