Refactor of how concurrency is handled.
The main difference now is that the ConfigWatcher base class now has a
"work_pool" for async jobs as well as a "thread_pool" for long-lived
tasks.

The Zookeeper discovery's start_watching() method is greatly simplified
thanks to this and now relies on kazoo's ChildrenWatch as a decorator
rather than calling it directly in a loop.

The Writer and Reporter classes are also simplified as they no longer
have to worry about when to launch threads.  They *do* have to worry
about the fact that each on_* method will be in a different thread.